fix: User model delete cascade configuration
This change adds the correct configuration to the `User` model to make
delete cascade work. This way, related objects to `User` can be properly
cascade deleted when appropriate.
